Mark non-deterministic test xfail
The test in run() itself works -- sometimes (i.e. the contract
creation via jsonrpc is functional at this point in time, see #66).
However in many cases a test fails due to the network not synchronizing
well. Marking it therefore xfail for now. Should be adressed at some
point in time!